% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{progr.lmu-package}
\alias{progr.lmu}
\alias{progr.lmu-package}
\title{progr.lmu: LMU Programming R Course}
\description{
Style guide and linting for the "Programming R" course at LMU.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mb706/lmu.progr}
  \item Report bugs at \url{https://github.com/mb706/lmu.progr/issues}
}

}
\author{
\strong{Maintainer}: Martin Binder \email{things.developer@mb706.com}

}
